[lib] Hide listener container struct

+++ lib/include/internal/hash.h
@@ -7,7 +7,7 @@
 
 HT_DECLS_BEGIN
 
-static uintptr_t djb2_hash(const char *str)
+static uint32_t djb2_hash(const char *str)
 {
     uint32_t hash = 5381;
     int c;
